requirements.txt
python-dotenv==1.0.1
transformers
torch
spaces
gradio